Transpiler: bare literal dim in new dynamic[<dim>]
`LOCAL aFoo[5]` was emitting `new dynamic[(int)(5)]` — the cast is
there for runtime expressions (`new dynamic[(int)(LASTREP)]` where
LASTREP is a #define resolving to a decimal) but reads as noise when
the dim is a long integer literal. Added hb_csEmitArrayDim helper
that emits a bare integer for HB_ET_NUMERIC/HB_ET_LONG and the
`(int)(...)` wrap for anything else; replaced the four allocation
sites (LOCAL, sub-LOCAL, PUBLIC, file-STATIC) with the helper.

Snapshot drift: test51 (`new dynamic[3]`) and test70 (`new dynamic[5]`
/ `[2]` / `[3]` for the dim'd-LOCAL/STATIC test). Both correct C#;
both still execute the same way at runtime.

+++ src/transpiler/tests/csout/test51.cs
@@ -25,7 +25,7 @@ public static partial class Program
     public static dynamic[] aFlag;
     public static void Main(string[] args)
     {
-        aFlag = new dynamic[(int)(3)];
+        aFlag = new dynamic[3];
 
         aFlag[0] = 10;
         aFlag[1] = 20;
